# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daybuilt API
 * OpenAPI spec version: 0.3.0
 */
export interface HealthStatus {
  status: string;
}

export type LeadInputSource =
  (typeof LeadInputSource)[keyof typeof LeadInputSource];

export const LeadInputSource = {
  contact: "contact",
  estimator: "estimator",
  "design-studio": "design-studio",
} as const;

export interface LeadInput {
  /** @minLength 2 */
  name: string;
  email: string;
  phone?: string;
  source: LeadInputSource;
  roomType?: string;
  roomSize?: string;
  builtInType?: string;
  materials?: string;
  budgetMin?: number;
  budgetMax?: number;
  description?: string;
  imageUrl?: string;
  style?: string;
  colorTone?: string;
  keepLayout?: string;
  timeline?: string;
  projectType?: string;
  /**
   * @minLength 32
   * @maxLength 200
   */
  designRequestId?: string;
  /** @minLength 32 */
  uploadProof?: string;
}

export interface LeadResponse {
  success: boolean;
  message: string;
  id: number;
  /** One-time anonymous access token for the design-studio AI generation endpoints. Returned only for source=design-studio and never retrievable again. Send it in the x-design-token header.
   */
  accessToken?: string;
}

export interface LeadRecord {
  id: number;
  name: string;
  email: string;
  phone?: string | null;
  source: string;
  status: string;
  roomType?: string | null;
  roomSize?: string | null;
  builtInType?: string | null;
  materials?: string | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  description?: string | null;
  imageUrl?: string | null;
  style?: string | null;
  colorTone?: string | null;
  keepLayout?: string | null;
  timeline?: string | null;
  projectType?: string | null;
  notes?: string | null;
  createdAt: string;
  /** pending | processing | completed | failed */
  aiStatus?: string | null;
  aiPromptSummary?: string | null;
  aiProvider?: string | null;
  aiModel?: string | null;
  aiAttempts?: number | null;
  aiError?: string | null;
  /** Admin-only diagnostic detail */
  aiErrorDetail?: string | null;
  aiStartedAt?: string | null;
  aiCompletedAt?: string | null;
  aiUpdatedAt?: string | null;
  hasGeneratedImage?: boolean | null;
  /** Protected API URL for the generated render */
  generatedImageUrl?: string | null;
  /** Protected API URL for the customer's uploaded photo */
  sourceImageUrl?: string | null;
}

export interface LeadsListResponse {
  leads: LeadRecord[];
}

export type UpdateLeadStatusInputStatus =
  (typeof UpdateLeadStatusInputStatus)[keyof typeof UpdateLeadStatusInputStatus];

export const UpdateLeadStatusInputStatus = {
  new: "new",
  contacted: "contacted",
  quoted: "quoted",
  closed: "closed",
} as const;

export interface UpdateLeadStatusInput {
  status: UpdateLeadStatusInputStatus;
  notes?: string;
}

/**
 * Customer-safe echo of the selections used to build the design.
 */
export interface GenerationSelections {
  roomType?: string | null;
  roomSize?: string | null;
  builtInType?: string | null;
  style?: string | null;
  colorTone?: string | null;
  keepLayout?: string | null;
  timeline?: string | null;
  budgetMin?: number | null;
  budgetMax?: number | null;
  description?: string | null;
}

export type GenerationStatusResponseStatus =
  (typeof GenerationStatusResponseStatus)[keyof typeof GenerationStatusResponseStatus];

export const GenerationStatusResponseStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

/**
 * Customer-facing generation state. Never contains the exact prompt, raw object paths, token hashes, or admin diagnostics.

 */
export interface GenerationStatusResponse {
  leadId: number;
  status: GenerationStatusResponseStatus;
  conceptSummary?: string | null;
  generatedImageUrl?: string | null;
  sourceImageUrl?: string | null;
  error?: string | null;
  attempts: number;
  startedAt?: string | null;
  completedAt?: string | null;
  updatedAt?: string | null;
  selections: GenerationSelections;
}

export type AdminGenerationStatusResponseStatus =
  (typeof AdminGenerationStatusResponseStatus)[keyof typeof AdminGenerationStatusResponseStatus];

export const AdminGenerationStatusResponseStatus = {
  pending: "pending",
  processing: "processing",
  completed: "completed",
  failed: "failed",
} as const;

/**
 * Admin view of generation state, including diagnostics.
 */
export interface AdminGenerationStatusResponse {
  leadId: number;
  status: AdminGenerationStatusResponseStatus;
  conceptSummary?: string | null;
  generatedImageUrl?: string | null;
  sourceImageUrl?: string | null;
  error?: string | null;
  errorDetail?: string | null;
  provider?: string | null;
  model?: string | null;
  attempts: number;
  startedAt?: string | null;
  completedAt?: string | null;
  updatedAt?: string | null;
}

export type UploadUrlRequestContentType =
  (typeof UploadUrlRequestContentType)[keyof typeof UploadUrlRequestContentType];

export const UploadUrlRequestContentType = {
  "image/jpeg": "image/jpeg",
  "image/jpg": "image/jpg",
  "image/png": "image/png",
  "image/webp": "image/webp",
} as const;

export interface UploadUrlRequest {
  /** @minLength 1 */
  name: string;
  /**
   * @minimum 1
   * @maximum 10485760
   */
  size: number;
  contentType: UploadUrlRequestContentType;
}

export interface UploadUrlResponse {
  uploadURL: string;
  objectPath: string;
  uploadToken: string;
  uploadProof: string;
}

export interface ErrorResponse {
  error: string;
}