# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.3.0
  description: Daybuilt API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads/{id}/generation:
    post:
      operationId: startLeadGeneration
      tags: [generation]
      summary: Start AI design generation for a design-studio lead
      description: >
        Starts the AI room render for a design-studio lead. Requires the
        one-time customer access token issued by POST /leads, sent in the
        x-design-token header. Generation runs inside the request lifecycle;
        the response reflects the final persisted state.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-design-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generation state (already completed, in progress, or finished now)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationStatusResponse"
        "400":
          description: Lead is not eligible for generation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Generation already in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationStatusResponse"
        "503":
          description: AI generation is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    get:
      operationId: getLeadGeneration
      tags: [generation]
      summary: Get AI design generation state
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-design-token
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generation state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GenerationStatusResponse"
        "401":
          description: Missing or invalid access token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/generation/retry:
    post:
      operationId: retryLeadGeneration
      tags: [generation]
      summary: Retry a failed AI generation (admin)
      description: >
        Admin-only retry. Allowed when the generation failed, or when a
        processing run is stale (older than the generation timeout).
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: x-admin-key
          in: header
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Generation state after the retry attempt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGenerationStatusResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Lead not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Generation is not in a retryable state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminGenerationStatusResponse"
        "503":
          description: Admin access or AI generation is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /leads/{id}/images/{kind}:
    get:
      operationId: getLeadImage
      tags: [generation]
      summary: Serve a protected lead image
      description: >
        Streams the customer's uploaded room photo (source) or the AI generated
        render (generated). Requires either the customer x-design-token or the
        admin x-admin-key header.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: kind
          in: path
          required: true
          schema:
            type: string
            enum: [source, generated]
        - name: x-design-token
          in: header
          required: false
          schema:
            type: string
        - name: x-admin-key
          in: header
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Image content
          content:
            image/png:
              schema:
                type: string
                format: binary
            image/jpeg:
              schema:
                type: string
                format: binary
            image/webp:
              schema:
                type: string
                format: binary
        "401":
          description: Missing or invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Image not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    GenerationSelections:
      type: object
      description: Customer-safe echo of the selections used to build the design.
      properties:
        roomType:
          type: string
          nullable: true
        roomSize:
          type: string
          nullable: true
        builtInType:
          type: string
          nullable: true
        style:
          type: string
          nullable: true
        colorTone:
          type: string
          nullable: true
        keepLayout:
          type: string
          nullable: true
        timeline:
          type: string
          nullable: true
        budgetMin:
          type: integer
          nullable: true
        budgetMax:
          type: integer
          nullable: true
        description:
          type: string
          nullable: true
    GenerationStatusResponse:
      type: object
      description: >
        Customer-facing generation state. Never contains the exact prompt,
        raw object paths, token hashes, or admin diagnostics.
      properties:
        leadId:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, failed]
        conceptSummary:
          type: string
          nullable: true
        generatedImageUrl:
          type: string
          nullable: true
        sourceImageUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        attempts:
          type: integer
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        selections:
          $ref: "#/components/schemas/GenerationSelections"
      required:
        - leadId
        - status
        - attempts
        - selections
    AdminGenerationStatusResponse:
      type: object
      description: Admin view of generation state, including diagnostics.
      properties:
        leadId:
          type: integer
        status:
          type: string
          enum: [pending, processing, completed, failed]
        conceptSummary:
          type: string
          nullable: true
        generatedImageUrl:
          type: string
          nullable: true
        sourceImageUrl:
          type: string
          nullable: true
        error:
          type: string
          nullable: true
        errorDetail:
          type: string
          nullable: true
        provider:
          type: string
          nullable: true
        model:
          type: string
          nullable: true
        attempts:
          type: integer
        startedAt:
          type: string
          format: date-time
          nullable: true
        completedAt:
          type: string
          format: date-time
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - leadId
        - status
        - attempts
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error