# lib/api-zod/src/generated/types/generationStatusResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Daybuilt API
 * OpenAPI spec version: 0.3.0
 */
import type { GenerationSelections } from "./generationSelections";
import type { GenerationStatusResponseStatus } from "./generationStatusResponseStatus";

/**
 * Customer-facing generation state. Never contains the exact prompt, raw object paths, token hashes, or admin diagnostics.

 */
export interface GenerationStatusResponse {
  leadId: number;
  status: GenerationStatusResponseStatus;
  conceptSummary?: string | null;
  generatedImageUrl?: string | null;
  sourceImageUrl?: string | null;
  error?: string | null;
  attempts: number;
  startedAt?: Date | null;
  completedAt?: Date | null;
  updatedAt?: Date | null;
  selections: GenerationSelections;
}